tts_api/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
accelerate==1.7.0
    # via transformers
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
boto3==1.38.18
    # via suno-bark
botocore==1.38.18
    # via
    #   boto3
    #   s3transfer
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   nltk
    #   uvicorn
einops==0.8.1
    # via encodec
encodec==0.1.1
    # via suno-bark
fastapi==0.115.12
    # via
    #   -r requirements.in
    #   fastapi-utils
fastapi-utils==0.8.0
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
funcy==2.0
    # via suno-bark
gunicorn==23.0.0
    # via
    #   -r requirements.in
    #   uvicorn-worker
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via
    #   -r requirements.in
    #   uvicorn
huggingface-hub==0.31.2
    # via
    #   accelerate
    #   suno-bark
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   requests
inotify==0.2.10
    # via -r requirements.in
jinja2==3.1.6
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.0
    # via nltk
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
mypy-extensions==1.1.0
    # via typing-inspect
networkx==3.4.2
    # via torch
nltk==3.9.1
    # via -r requirements.in
nose==1.3.7
    # via inotify
numpy==2.2.6
    # via
    #   accelerate
    #   encodec
    #   scipy
    #   suno-bark
    #   transformers
nvidia-ml-py==12.575.51
    # via pynvml
packaging==25.0
    # via
    #   accelerate
    #   gunicorn
    #   huggingface-hub
    #   transformers
psutil==5.9.8
    # via
    #   accelerate
    #   fastapi-utils
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-utils
pydantic-core==2.33.2
    # via pydantic
pynvml==12.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.0
    # via uvicorn
python-json-logger==3.3.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via
    #   nltk
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
s3transfer==0.12.0
    # via boto3
safetensors==0.5.3
    # via
    #   accelerate
    #   transformers
scipy==1.15.3
    # via
    #   -r requirements.in
    #   suno-bark
setuptools==80.7.1
    # via torch
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
    # via -r requirements.in
sympy==1.14.0
    # via torch
tokenizers==0.21.1
    # via
    #   suno-bark
    #   transformers
torch==2.7.0
    # via
    #   accelerate
    #   encodec
    #   suno-bark
    #   torchaudio
    #   transformers
torchaudio==2.7.0
    # via encodec
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   suno-bark
    #   transformers
transformers==4.51.3
    # via
    #   -r requirements.in
    #   suno-bark
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via -r requirements.in
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via
    #   botocore
    #   requests
uvicorn==0.34.2
    # via
    #   -r requirements.in
    #   uvicorn-worker
uvicorn-worker==0.3.0
    # via -r requirements.in
uvloop==0.21.0
    # via
    #   -r requirements.in
    #   uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
